+++ unikmer/cmd/count.go
@@ -68,7 +68,7 @@ var countCmd = &cobra.Command{
 		if !isStdout(outFile) {
 			outFile += extDataFile
 		}
-		outfh, gw, w, err := outStream(outFile, opt.Compress)
+		outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
 		checkError(err)
 		defer func() {
 			outfh.Flush()