+++ unikmer/cmd/diff.go
@@ -103,7 +103,7 @@ Tips:
 				if !isStdout(outFile) {
 					outFile += extDataFile
 				}
-				outfh, gw, w, err := outStream(outFile, opt.Compress)
+				outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
 				checkError(err)
 				defer func() {
 					outfh.Flush()
@@ -230,7 +230,7 @@ Tips:
 			if !isStdout(outFile) {
 				outFile += extDataFile
 			}
-			outfh, gw, w, err := outStream(outFile, opt.Compress)
+			outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
 			checkError(err)
 			defer func() {
 				outfh.Flush()
@@ -477,7 +477,7 @@ Tips:
 		if !isStdout(outFile) {
 			outFile += extDataFile
 		}
-		outfh, gw, w, err := outStream(outFile, opt.Compress)
+		outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
 		checkError(err)
 		defer func() {
 			outfh.Flush()