+++ unikmer/cmd/inter.go
@@ -96,7 +96,7 @@ var interCmd = &cobra.Command{
 					if !isStdout(outFile) {
 						outFile += extDataFile
 					}
-					outfh, gw, w, err := outStream(outFile, opt.Compress)
+					outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
 					checkError(err)
 					defer func() {
 						outfh.Flush()
@@ -260,7 +260,7 @@ var interCmd = &cobra.Command{
 		if !isStdout(outFile) {
 			outFile += extDataFile
 		}
-		outfh, gw, w, err := outStream(outFile, opt.Compress)
+		outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
 		checkError(err)
 		defer func() {
 			outfh.Flush()